+++ man-roxygen/section_rules.R
@@ -1,4 +1,4 @@
-#' <% p <- getFromNamespace("prim", "anvil")(primitive_id); implemented <- Filter(function(r) !is.null(p$rules[[r]]), getFromNamespace("globals", "anvil")$interpretation_rules) %>
+#' <% p <- primitive_env[[primitive_id]]; implemented <- Filter(function(r) !is.null(p[[r]]), globals$interpretation_rules) %>
 #' <% if (length(implemented) > 0) { %>
 #' @section Implemented Rules:
 #' <%= paste0("- `", implemented, "`", collapse = "\n#' ") %>